# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GymApp API
  version: "1.0.0"
  description: |
    Endpoints HTTP de GymApp (`app/api/**/route.ts`, Next.js App Router).

    **Cómo se autentica cada endpoint** (ver `security` en cada operación):
    - `sesion`: cookies de sesión de Supabase que el navegador envía solo (`@supabase/ssr`). Se usa desde las páginas de la app.
    - `bearerUsuario`: `Authorization: Bearer <access_token de Supabase>` del usuario.
    - `cronSecret`: `Authorization: Bearer <CRON_SECRET>` para Vercel Cron (env var `CRON_SECRET`).
    - `security: []`: endpoint **público**, sin autenticación.

    "Requiere super_admin" significa que el perfil del usuario (`profiles.role`) debe ser `super_admin`;
    "admin completo" significa además ser el dueño (`gym_config.owner_email`) o `NEXT_PUBLIC_ADMIN_EMAIL` (`lib/admin-level.ts`).

    **Límite de peticiones** (`lib/middleware/rate-limit.ts`): cuando aplica, responde `429` con `Retry-After` y
    `X-RateLimit-*`. Si el RPC de rate limit falla, la petición pasa (fail-open).

    Los endpoints marcados con ⚠️ tienen hallazgos pendientes documentados en `openspec/ROADMAP.md`.

    Este archivo se valida en `__tests__/api-docs.test.ts`: cada `route.ts` debe estar documentado aquí y viceversa.
servers:
  - url: http://localhost:3000
    description: Local (`npm run dev`)
...
paths:
...
  /api/notificaciones:
    get:
      tags: [notificaciones]
      operationId: ejecutarNotificacionesCron
      summary: Ejecuta las notificaciones habilitadas (Vercel Cron)
      description: |
        Invocado por Vercel Cron (`vercel.json`, diario a las 04:00 UTC = medianoche VET) con
        `Authorization: Bearer <CRON_SECRET>` que Vercel envía automáticamente. Misma lógica que el POST.
        Si `CRON_SECRET` no está definida en el entorno, el cron responde 401 (fail-closed).
        Al terminar envía un reporte en texto plano a `NEXT_PUBLIC_ADMIN_EMAIL` (solo esta vía del cron;
        las ejecuciones manuales no generan reporte).
      security:
        - cronSecret: []
        - bearerUsuario: []
      responses:
        "200":
          description: Resultado de la ejecución.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string, description: "Solo cuando no se ejecutó nada (deshabilitadas o sin configuraciones)." }
                  ejecutadas: { type: integer }
                  enviados: { type: integer }
                  errores: { type: integer }
                  reporte_enviado: { type: [boolean, "null"], description: "true si el reporte de corrida llegó al correo del super admin (null en guardas y en ejecuciones manuales)." }
                  reporte: { type: [string, "null"], description: "Texto plano del reporte de corrida." }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500":
          description: Error inesperado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: false }
                  error: { type: string }
    post:
      tags: [notificaciones]
      operationId: ejecutarNotificaciones
      summary: Ejecuta las notificaciones habilitadas (cron diario)
      description: |
        Recomendado: cron diario a medianoche. Si `gym_config.notifications_enabled` es falso o no hay tipos habilitados,
        no hace nada. Para cada tipo habilitado en `notificacion_config`, respeta su frecuencia y envía los correos
        (deudores, recordatorio de pago, resumen al dueño, estatus del sistema); registra en `notification_log`
        al final y solo si ningún tipo falló (la bitácora se purga dejando las 10 filas más recientes por tipo).
        Límite de 5 cada 20 min solo cuando llama un usuario (no con `CRON_SECRET`).

        Al terminar envía un reporte en texto plano a `NEXT_PUBLIC_ADMIN_EMAIL` con destinatarios, % de cobertura
        y errores por plantilla (solo cuando se invoca con `CRON_SECRET`; el POST de usuario no genera reporte).

        Si `CRON_SECRET` no está definida en el entorno, el cron responde 401 (fail-closed).
      security:
        - cronSecret: []
        - bearerUsuario: []
      responses:
        "200":
          description: Resultado de la ejecución.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string, description: "Solo cuando no se ejecutó nada (deshabilitadas o sin configuraciones)." }
                  ejecutadas: { type: integer }
                  enviados: { type: integer }
                  errores: { type: integer }
                  reporte_enviado: { type: [boolean, "null"], description: "true si el reporte de corrida llegó al correo del super admin (null en guardas y en ejecuciones manuales)." }
                  reporte: { type: [string, "null"], description: "Texto plano del reporte de corrida." }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500":
          description: Error inesperado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: false }
                  error: { type: string }
  /api/notificaciones/procesar:
    post:
      tags: [notificaciones]
      operationId: procesarNotificaciones
      summary: Ejecuta notificaciones manualmente ("Ejecutar ahora")
      description: |
        Requiere super_admin (token Bearer del usuario). Con `tipo`, ejecuta solo ese tipo; con `forzar: true`,
        ignora la frecuencia. Si las notificaciones están deshabilitadas o no hay configuraciones, devuelve ceros.
      security:
        - bearerUsuario: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                tipo:
                  type: string
                  enum: [miembros_deudores, recordatorio_pago, resumen_dueno, estatus_sistema]
                forzar: { type: boolean, default: false }
      responses:
        "200":
          description: Resultado de la ejecución.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ejecutadas: { type: integer }
                  enviados: { type: integer }
                  errores: { type: integer }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "500": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    sesion:
      type: apiKey
      in: cookie
      name: sb-<project-ref>-auth-token
      description: Cookies de sesión de Supabase (`@supabase/ssr`); el navegador las envía solo.
    bearerUsuario:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token de Supabase del usuario (`supabase.auth.getSession()`).
    cronSecret:
      type: http
      scheme: bearer
      description: Valor de la variable de entorno `CRON_SECRET`.
  responses:
    Error:
      description: Error con mensaje para el usuario (textos de `lib/messages.ts`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: Límite de peticiones superado.
      headers:
        Retry-After: { schema: { type: integer }, description: Segundos hasta poder reintentar. }
        X-RateLimit-Limit: { schema: { type: integer } }
        X-RateLimit-Remaining: { schema: { type: integer } }
        X-RateLimit-Reset: { schema: { type: integer }, description: Epoch en segundos. }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...